// src/app/feature/model/education.model.ts
export interface Education{
    id:number;
    employeeId: number;
    egucationGroupId: number;
    educationTypeId: number;
    gradeId: number;
    passingYear: string;
    result: string;
    scale: string;
    instituteId: number;
    foreignDegree: string;
    professionalDegree: string;
    lastEducation: string;
    remarks: string;
}